+++ src/components/MainLayout/FooterCompo.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import { Layout } from 'antd';
+const { Footer } = Layout;
+
+const FooterComp = () => {
+  return (
+    <Footer
+      style={{
+        textAlign: 'center',
+      }}
+    >
+      Ant Design ©2018 Created by Ant UED
+    </Footer>
+  );
+};
+
+export default FooterComp;
