+++ src/components/MainLayout/HeaderNaviCompo.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Breadcrumb } from 'antd';
+const HeaderNaviCompo = () => {
+  return (
+    <Breadcrumb
+      style={{
+        margin: '16px 0',
+      }}
+    >
+      <Breadcrumb.Item>User</Breadcrumb.Item>
+      <Breadcrumb.Item>Bill</Breadcrumb.Item>
+    </Breadcrumb>
+  );
+};
+
+export default HeaderNaviCompo;
