--- src/pages/MainLayout.js
@@ -1,90 +1,0 @@
-import {
-  DesktopOutlined,
-  PieChartOutlined,
-  UserOutlined,
-} from '@ant-design/icons';
-import { Breadcrumb, Layout, Menu } from 'antd';
-import React, { useState } from 'react';
-import Main from './Main';
-import Borad from './Board';
-import 'antd/dist/antd.css';
-
-const { Header, Content, Footer, Sider } = Layout;
-function getItem(label, key, icon, children) {
-  return {
-    key,
-    icon,
-    children,
-    label,
-  };
-}
-const items = [
-  getItem('메인', '1', <UserOutlined />),
-  getItem('게시판', '2', <PieChartOutlined />),
-  getItem('개인설정', '3', <DesktopOutlined />),
-];
-const MainLayout = () => {
-  const [collapsed, setCollapsed] = useState(false);
-  const [page, setPage] = useState('1');
-
-  const onClick = e => {
-    setPage(e.key);
-  };
-
-  return (
-    <Layout
-      style={{
-        minHeight: '100vh',
-      }}
-    >
-      <Sider
-        collapsible
-        collapsed={collapsed}
-        onCollapse={value => setCollapsed(value)}
-      >
-        <div className="logo" />
-        <Menu
-          theme="dark"
-          defaultSelectedKeys={page}
-          mode="inline"
-          items={items}
-          onClick={onClick}
-        />
-      </Sider>
-      <Layout className="site-layout">
-        <Header
-          className="site-layout-background"
-          style={{
-            padding: 0,
-          }}
-        />
-        <Content
-          style={{
-            margin: '16px',
-            padding: '16px',
-            backgroundColor: 'white',
-          }}
-        >
-          <Breadcrumb
-            style={{
-              margin: '16px 0',
-            }}
-          >
-            <Breadcrumb.Item>User</Breadcrumb.Item>
-            <Breadcrumb.Item>Bill</Breadcrumb.Item>
-          </Breadcrumb>
-          {page === '1' && <Main />}
-          {page === '2' && <Borad />}
-        </Content>
-        <Footer
-          style={{
-            textAlign: 'center',
-          }}
-        >
-          Ant Design ©2018 Created by Ant UED
-        </Footer>
-      </Layout>
-    </Layout>
-  );
-};
-export default MainLayout;
